allow tasks to request dedicated stack space when created
never copy over the root stack:
this is a hybrid approach to COPY_STACK where the root task is never
moved or copied, and all other task stacks are layered into the same
memory area

use ucontext_t where it is available (aka linux)

always use makecontext when starting a Task and creating a new stack

identify all of the available task context creation algorithms with HAVE_ flags for better clarity and simplicity

MINSTKSZ on linux is absurdly low. enforce some sanity on it.